backend_alan/requirements.txt
livekit-agents
livekit-plugins-google
google-generativeai
python-dotenv
openai
